(src/include) Adjusting code
- Removing references to vroom
- Fixing filenames on the license

+++ src/pgr_pickDeliver/pickDeliverEuclidean.c
@@ -137,7 +137,7 @@ _vrp_pgr_pickdelivereuclidean(PG_FUNCTION_ARGS) {
             nulls[i] = false;
         }
 
-        values[0] = Int32GetDatum(funcctx->call_cntr + 1);
+        values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
         values[1] = Int32GetDatum(result_tuples[call_cntr].vehicle_seq);
         values[2] = Int64GetDatum(result_tuples[call_cntr].vehicle_id);
         values[3] = Int32GetDatum(result_tuples[call_cntr].stop_seq);
