+++ src/pickDeliver/pickDeliver.c
@@ -168,7 +168,7 @@ _vrp_pickdeliver(PG_FUNCTION_ARGS) {
     }
 
 
-    values[0] = Int32GetDatum(funcctx->call_cntr + 1);
+    values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
     values[1] = Int32GetDatum(result_tuples[call_cntr].vehicle_seq);
     values[2] = Int64GetDatum(result_tuples[call_cntr].vehicle_id);
     values[3] = Int32GetDatum(result_tuples[call_cntr].stop_seq);
@@ -261,7 +261,7 @@ _vrp_pickdeliverraw(PG_FUNCTION_ARGS) {
     }
 
 
-    values[0] = Int32GetDatum(funcctx->call_cntr + 1);
+    values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
     values[1] = Int32GetDatum(result_tuples[call_cntr].vehicle_seq);
     values[2] = Int64GetDatum(result_tuples[call_cntr].vehicle_id);
     values[3] = Int32GetDatum(result_tuples[call_cntr].stop_seq);
